backend/requirements.txt
alembic==1.8.1
apispec==5.2.2
asgiref==3.5.2
asyncpg==0.26.0
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
colorama==0.4.5
dependency-injector==4.39.1
Flask==2.1.3
Flask-JSON==0.3.4
Flask-JWT-Extended==4.4.2
flask-smorest==0.38.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
packaging==21.3
passlib==1.7.4
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.39
toml==0.10.2
webargs==8.2.0
Werkzeug==2.1.2
